Update documentation generation
Documentation should be easier to update, and using different pipeline
for this step seems to be a better choice than the old one.

I added `ember-cli-yuidoc` and an npm script to allow easy generation.

As for the theme, I choose to use `yuidoc-ember-cli-theme` used for the
`ember-cli` [API documentation](https://ember-cli.com/api/).

I also think there is no need to split code and documentation. They can
both coexist in the same branch, and documentation will be versioned with
the code. The only side-effect is to configure this repository to use
the `docs` directory for GitHub pages.