Mark raven-shim as an ES module
Ember's AMD loader is tolerant if you don't do this, but other build systems are not. I'm experimenting with getting a large number of addons compatible with other build tools (like Webpack and Parcel). In my testing this bug came up.

The bug is that consumers can't tell whether it's already a module that has a default export vs some random object that needs to get wrapped in another `{ default: ... }` to make it compatibility with consuming modules.